Refactoring on Caregiver tests
