+++ packages/yourloops/test/integration/pages/home/home.caregiver.spec.tsx
@@ -25,26 +25,30 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { loggedInUserId, mockAuth0Hook } from '../../mock/auth0.hook.mock'
-import { mockNotificationAPI } from '../../mock/notification.api.mock'
-import { mockDirectShareApi, removeDirectShareMock } from '../../mock/direct-share.api.mock'
-import { patient1Info, patient2Info } from '../../data/patient.api.data'
-import { mockTeamAPI } from '../../mock/team.api.mock'
-import { checkCaregiverLayout } from '../../assert/layout.assert'
-import { renderPage } from '../../utils/render'
-import { act, screen, within } from '@testing-library/react'
-import userEvent from '@testing-library/user-event'
-import DirectShareApi from '../../../../lib/share/direct-share.api'
+import { act, screen } from '@testing-library/react'
+import moment from 'moment-timezone'
 import { UserRole } from '../../../../lib/auth/models/enums/user-role.enum'
-import { mockUserApi } from '../../mock/user.api.mock'
-import { mockPatientApiForCaregivers } from '../../mock/patient.api.mock'
 import PatientApi from '../../../../lib/patient/patient.api'
+import DirectShareApi from '../../../../lib/share/direct-share.api'
+import { checkCaregiverLayout } from '../../assert/layout.assert'
 import { checkPatientListHeaderCaregiver } from '../../assert/patient-list.assert'
 import { buildPatient } from '../../data/patient-builder.data'
-import moment from 'moment-timezone'
-import { mockDblCommunicationApi } from '../../mock/dbl-communication.api'
-import { mockErrorApi } from '../../mock/error.api.mock'
 import { mockAnalyticsApi } from '../../mock/analytics.api.mock'
+import { mockAuth0Hook } from '../../mock/auth0.hook.mock'
+import { mockDblCommunicationApi } from '../../mock/dbl-communication.api'
+import { mockDirectShareApi } from '../../mock/direct-share.api.mock'
+import { mockErrorApi } from '../../mock/error.api.mock'
+import { mockNotificationAPI } from '../../mock/notification.api.mock'
+import { mockPatientApiForCaregivers } from '../../mock/patient.api.mock'
+import { mockTeamAPI } from '../../mock/team.api.mock'
+import { mockUserApi } from '../../mock/user.api.mock'
+import {
+  testPatientListColumnsCaregiver,
+  testPatientListSearchCaregiver,
+  testRemovePatientCaregiver,
+  testRemovePatientErrorCaregiver
+} from '../../use-cases/patient-list-management'
+import { renderPage } from '../../utils/render'
 
 describe('Caregiver home page', () => {
   const firstName = 'Eric'
@@ -132,26 +136,7 @@ describe('Caregiver home page', () => {
 
     const lastDataUploadDate = moment.tz(patient1.medicalData.range.endDate, new Intl.DateTimeFormat().resolvedOptions().timeZone).format('lll')
 
-    // Checking that all patients are displayed
-    const dataGridRow = screen.getByTestId('current-patient-list-grid')
-    expect(within(dataGridRow).getAllByRole('row')).toHaveLength(4)
-    expect(dataGridRow).toHaveTextContent(`PatientDate of birthTIRBelow rangeLast data updateActionsFlag patient Akim@embett.comEmbett AkimJan 20, 20100%0%Jun 22, 2023 7:02 AMFlag patient alain@provist.comProvist AlainJan 20, 20100%0%Jun 22, 2023 7:02 AMFlag patient annie@versaire.comVersaire AnnieMay 25, 20150%0%Jun 22, 2023 7:02 AM`)
-
-    const searchPatient = screen.getByPlaceholderText('Search for a patient...')
-
-    // Searching by birthdate only
-    await userEvent.type(searchPatient, '20/01/2010')
-    expect(dataGridRow).toHaveTextContent(`PatientDate of birthTIRBelow rangeLast data updateActionsFlag patient Akim@embett.comEmbett AkimJan 20, 20100%0%Jun 22, 2023 7:02 AMFlag patient alain@provist.comProvist AlainJan 20, 20100%0%Jun 22, 2023 7:02 AM`)
-    await userEvent.clear(searchPatient)
-
-    // Searching by birthdate and first name
-    await userEvent.type(searchPatient, '20/01/2010 Aki')
-    expect(dataGridRow).toHaveTextContent(`PatientDate of birthTIRBelow rangeLast data updateActionsFlag patient Akim@embett.comEmbett AkimJan 20, 20100%0%${lastDataUploadDate}`)
-    await userEvent.clear(searchPatient)
-
-    // Searching by birthdate and last name
-    await userEvent.type(searchPatient, '20/01/2010provi')
-    expect(dataGridRow).toHaveTextContent('PatientDate of birthTIRBelow rangeLast data updateActionsFlag patient alain@provist.comProvist AlainJan 20, 20100%0%Jun 22, 2023 7:02 AM')
+    await testPatientListSearchCaregiver(lastDataUploadDate)
   })
 
   it('should display a list of patients and allow to remove one of them', async () => {
@@ -159,48 +144,7 @@ describe('Caregiver home page', () => {
       renderPage('/')
     })
 
-    await checkCaregiverLayout(`${lastName} ${firstName}`)
-    checkPatientListHeaderCaregiver()
-
-    const patientTableBody = screen.getByTestId('current-patient-list-grid')
-    expect(within(patientTableBody).getAllByRole('row')).toHaveLength(5)
-    expect(patientTableBody).toHaveTextContent('PatientDate of birthTIRBelow rangeLast data updateActionsFlag patient patient1@diabeloop.frGroby Patient1Jan 1, 1980N/AFlag patient pending-patient@diabeloop.frPatient PendingJan 1, 1980N/AFlag patient patient2@diabeloop.frRouis Patient2Jan 1, 1980N/AFlag patient patient3@diabeloop.frSrairi Patient3Jan 1, 1980N/A')
-    const removePatientButton = screen.getByRole('button', { name: `Remove patient ${patient2Info.profile.email}` })
-    expect(removePatientButton).toBeVisible()
-
-    await userEvent.click(removePatientButton)
-
-    const removePatientDialog = screen.getByRole('dialog')
-    expect(removePatientDialog).toBeVisible()
-
-    const removePatientDialogTitle = within(removePatientDialog).getByText('Remove patient Patient2 Rouis')
-    expect(removePatientDialogTitle).toBeVisible()
-
-    expect(removePatientDialog).toHaveTextContent('Are you sure you want to remove Patient2 Rouis?')
-
-    const removePatientDialogCancelButton = within(removePatientDialog).getByText('Cancel')
-    expect(removePatientDialogCancelButton).toBeVisible()
-
-    const removePatientDialogConfirmButton = within(removePatientDialog).getByRole('button', { name: 'Remove patient' })
-    expect(removePatientDialogConfirmButton).toBeVisible()
-
-    await userEvent.click(removePatientDialogCancelButton)
-
-    expect(removePatientDialog).not.toBeInTheDocument()
-
-    await userEvent.click(removePatientButton)
-
-    const removePatientDialog2 = screen.getByRole('dialog')
-    expect(removePatientDialog2).toBeVisible()
-
-    const removePatientDialog2ConfirmButton = within(removePatientDialog2).getByRole('button', { name: 'Remove patient' })
-
-    await userEvent.click(removePatientDialog2ConfirmButton)
-
-    expect(removeDirectShareMock).toHaveBeenCalledWith(patient2Info.userid, loggedInUserId)
-    expect(jest.spyOn(PatientApi, 'getPatientsForCaregivers').mockResolvedValue([patient1Info])).toHaveBeenCalledTimes(2)
-    expect(screen.queryByTestId('remove-direct-share-dialog')).toBeFalsy()
-    expect(screen.getByTestId('alert-snackbar')).toHaveTextContent('You no longer have access to your patient\'s data.')
+    await testRemovePatientCaregiver(lastName, firstName)
   })
 
   it('should display an error message if patient removal failed', async () => {
@@ -210,19 +154,7 @@ describe('Caregiver home page', () => {
       renderPage('/')
     })
 
-    const removeButton = screen.getByRole('button', { name: `Remove patient ${patient2Info.profile.email}` })
-
-    await userEvent.click(removeButton)
-
-    const removeDialog = screen.getByRole('dialog')
-
-    const confirmRemoveButton = within(removeDialog).getByRole('button', { name: 'Remove patient' })
-
-    await userEvent.click(confirmRemoveButton)
-
-    expect(removeDirectShareMock).toHaveBeenCalledWith(patient2Info.userid, loggedInUserId)
-    expect(screen.getByTestId('remove-direct-share-dialog')).toBeVisible()
-    expect(screen.getByTestId('alert-snackbar')).toHaveTextContent('Impossible to remove patient. Please try again later.')
+    await testRemovePatientErrorCaregiver()
   })
 
   it('should display the right columns for caregivers and be able to change them', async () => {
@@ -230,56 +162,6 @@ describe('Caregiver home page', () => {
       renderPage('/')
     })
 
-    const columnSettingsButton = screen.getByRole('button', { name: 'Change columns settings' })
-
-    checkPatientListHeaderCaregiver()
-    expect(screen.getByRole('columnheader', { name: 'Patient' })).toBeVisible()
-    expect(screen.queryByRole('columnheader', { name: 'Age' })).not.toBeInTheDocument()
-    expect(screen.getByRole('columnheader', { name: 'Date of birth' })).toBeVisible()
-    expect(screen.queryByRole('columnheader', { name: 'Gender' })).not.toBeInTheDocument()
-    expect(screen.queryByRole('columnheader', { name: 'System' })).not.toBeInTheDocument()
-    expect(screen.getByRole('columnheader', { name: 'Last data update' })).toBeVisible()
-    expect(screen.getByRole('columnheader', { name: 'Actions' })).toBeVisible()
-
-    await userEvent.click(columnSettingsButton)
-
-    const columnSettingsPopover = screen.getByRole('presentation')
-    const applyButton = within(columnSettingsPopover).getByRole('button', { name: 'Apply' })
-    expect(columnSettingsPopover).toBeVisible()
-    expect(within(columnSettingsPopover).getByText('Show column')).toBeVisible()
-
-    const patientToggle = within(within(columnSettingsPopover).getByLabelText('Patient')).getByRole('switch')
-    const ageToggle = within(within(columnSettingsPopover).getByLabelText('Age')).getByRole('switch')
-    const dateOfBirthToggle = within(within(columnSettingsPopover).getByLabelText('Date of birth')).getByRole('switch')
-    const genderToggle = within(within(columnSettingsPopover).getByLabelText('Gender')).getByRole('switch')
-    const systemToggle = within(within(columnSettingsPopover).getByLabelText('System')).getByRole('switch')
-    const lastUpdateToggle = within(within(columnSettingsPopover).getByLabelText('Last data update')).getByRole('switch')
-    expect(patientToggle).toHaveProperty('checked', true)
-    expect(patientToggle).toHaveProperty('disabled', true)
-    expect(ageToggle).toHaveProperty('checked', false)
-    expect(dateOfBirthToggle).toHaveProperty('checked', true)
-    expect(genderToggle).toHaveProperty('checked', false)
-    expect(systemToggle).toHaveProperty('checked', false)
-    expect(lastUpdateToggle).toHaveProperty('checked', true)
-
-    const disabledToggle = screen.getByLabelText('This column cannot be removed')
-    await userEvent.hover(disabledToggle)
-    expect(await screen.findByText('This column cannot be removed')).toBeVisible()
-
-    await userEvent.click(ageToggle)
-    await userEvent.click(lastUpdateToggle)
-    expect(ageToggle).toHaveProperty('checked', true)
-    expect(lastUpdateToggle).toHaveProperty('checked', false)
-
-    await userEvent.click(applyButton)
-
-    expect(columnSettingsPopover).not.toBeVisible()
-    expect(screen.getByRole('columnheader', { name: 'Patient' })).toBeVisible()
-    expect(screen.getByRole('columnheader', { name: 'Age' })).toBeVisible()
-    expect(screen.getByRole('columnheader', { name: 'Date of birth' })).toBeVisible()
-    expect(screen.queryByRole('columnheader', { name: 'Gender' })).not.toBeInTheDocument()
-    expect(screen.queryByRole('columnheader', { name: 'System' })).not.toBeInTheDocument()
-    expect(screen.queryByRole('columnheader', { name: 'Last data update' })).not.toBeInTheDocument()
-    expect(screen.getByRole('columnheader', { name: 'Actions' })).toBeVisible()
+    await testPatientListColumnsCaregiver()
   })
 })
